updated styling of anchor films to show ratings as stars and updated poster styling
