+++ src/client/components/UserComparison.tsx
@@ -3,7 +3,7 @@ import RatingDistributionHistogram from "./RatingDistributionHistogram";
 import TasteCompatibility from "./TasteCompatibility";
 import { MoviesInCommonData } from "../types";
 import { useComparison } from "../hooks/useComparison";
-
+import StarRating from "./StarRating";
 interface Rating {
   rating: number;
   count: number;
@@ -20,11 +20,7 @@ interface UserData {
 }
 
 const UserComparison = () => {
-  const {
-    usernames,
-    getUserComplete,
-    getMoviesInCommon,
-  } = useComparison();
+  const { usernames, getUserComplete, getMoviesInCommon } = useComparison();
   const [selectedUser1, setSelectedUser1] = useState<string>("");
   const [selectedUser2, setSelectedUser2] = useState<string>("");
   const [user1Data, setUser1Data] = useState<UserData | null>(null);
@@ -57,7 +53,7 @@ const UserComparison = () => {
       }
     } catch (err) {
       setError(
-        err instanceof Error ? err.message : "Failed to load movies in common"
+        err instanceof Error ? err.message : "Failed to load movies in common",
       );
     } finally {
       setLoadingMovies(false);
@@ -91,7 +87,7 @@ const UserComparison = () => {
       }
     } catch (err) {
       setError(
-        err instanceof Error ? err.message : "Failed to load user ratings"
+        err instanceof Error ? err.message : "Failed to load user ratings",
       );
     } finally {
       setLoading(false);
@@ -133,45 +129,11 @@ const UserComparison = () => {
 
     const totalWeightedRating = userData.ratings.reduce(
       (sum, r) => sum + r.rating * r.count,
-      0
+      0,
     );
     const totalCount = getTotalRatings(userData);
 
     return totalCount > 0 ? totalWeightedRating / totalCount : 0;
-  };
-
-  const StarRating = ({ rating }: { rating: number }) => {
-    const fullStars = Math.floor(rating);
-    const hasHalfStar = rating % 1 !== 0;
-    const emptyStars = 5 - fullStars - (hasHalfStar ? 1 : 0);
-
-    return (
-      <div className="flex items-center">
-        {Array.from({ length: fullStars }, (_, i) => (
-          <span key={`full-${i}`} className="text-letterboxd-accent text-lg">
-            ★
-          </span>
-        ))}
-
-        {hasHalfStar && (
-          <div key="half" className="relative inline-block text-lg">
-            <span className="text-letterboxd-border">☆</span>
-            <span
-              className="absolute inset-0 text-letterboxd-accent overflow-hidden"
-              style={{ width: "50%" }}
-            >
-              ★
-            </span>
-          </div>
-        )}
-
-        {Array.from({ length: emptyStars }, (_, i) => (
-          <span key={`empty-${i}`} className="text-letterboxd-border text-lg">
-            ☆
-          </span>
-        ))}
-      </div>
-    );
   };
 
   return (
